LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-align-baseline-multicol-002-expected.txt










































FAIL .target > * 1 assert_equals:
<div data-offset-x="105"><span></span></div>
offsetLeft expected 105 but got 135
PASS .target > * 2
FAIL .target > * 3 assert_equals:
<div data-offset-x="35"><span></span></div>
offsetLeft expected 35 but got 10
FAIL .target > * 4 assert_equals:
<div class="multicol" data-offset-x="10">
    <div style="font-size: 10px;">
      <span></span><br><span></span><br><span></span>
    </div>
    <div style="font-size: 30px;">
      <span></span><br><span></span><br><span></span>
    </div>
    <div style="font-size: 20px;">
      <span></span><br><span></span><br><span></span>
    </div>
  </div>
offsetLeft expected 10 but got 172
FAIL .target > * 5 assert_equals:
<div data-offset-x="140"><span></span></div>
offsetLeft expected 140 but got 175
PASS .target > * 6
FAIL .target > * 7 assert_equals:
<div data-offset-x="35"><span></span></div>
offsetLeft expected 35 but got 120
FAIL .target > * 8 assert_equals:
<div class="multicol" data-offset-x="10">
    <div style="column-span: all; background: orange;">
      <span></span><br><span></span>
    </div>
    <div style="font-size: 10px;">
      <span></span><br><span></span><br><span></span>
    </div>
    <div style="font-size: 30px;">
      <span></span><br><span></span><br><span></span>
    </div>
    <div style="font-size: 20px;">
      <span></span><br><span></span><br><span></span>
    </div>
  </div>
offsetLeft expected 10 but got -20
FAIL .target > * 9 assert_equals:
<div data-offset-x="145"><span></span></div>
offsetLeft expected 145 but got 176
PASS .target > * 10
FAIL .target > * 11 assert_equals:
<div data-offset-x="30"><span></span></div>
offsetLeft expected 30 but got 31
PASS .target > * 12
